refactoring
class divid
